[Sema] Move `getDocCommentProvidingDecl` and `getCascadingDocComment` from `AST` to `Sema`
This allows use to re-use logic from Sema in those requests.

This commit just moves functions around and does not change any functionality.
